chore(common): De-coupling available models from lazy loader.
