--- archai/nlp/common/constants.py
@@ -1,16 +1,0 @@
-# Copyright (c) Microsoft Corporation.
-# Licensed under the MIT license.
-
-"""Constants definitions used within the NLP package.
-"""
-
-# Lazy loader-related
-LIBRARY_PATH = 'archai.nlp.models'
-
-# Models-related
-BATCH_SIZE = 1
-SEQ_LEN = 32
-
-# ONNX-related
-OMP_NUM_THREADS = 1
-OMP_WAIT_POLICY = 'ACTIVE'
