+++ archai/nlp/compression/onnx/onnx_utils/onnx_loader.py
@@ -8,7 +8,6 @@ import types
 from os import environ
 from typing import Any, Dict, Sized, Tuple
 
-from archai.nlp.common.constants import OMP_NUM_THREADS, OMP_WAIT_POLICY
 from archai.nlp.common.lazy_loader import load_model_from_checkpoint
 from archai.nlp.compression.onnx.onnx_utils.forward import (
     crit_forward_memformer_onnx, forward_gpt2_onnx, forward_memformer_onnx)
@@ -16,6 +15,10 @@ from archai.nlp.models.model_base import ArchaiModel
 from onnxruntime import (GraphOptimizationLevel, InferenceSession,
                          SessionOptions)
 from onnxruntime.transformers import quantize_helper
+
+# ONNX-loading constants
+OMP_NUM_THREADS = 1
+OMP_WAIT_POLICY = 'ACTIVE'
 
 # Constants available in onnxruntime
 # that enables performance optimization
